tests/wpt/tests/css/css-sizing/aspect-ratio/fractional-aspect-ratio.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('div')">
  <div style="width: 100px; aspect-ratio: 0.00025 / 0.0001;" data-expected-height=40></div>
  <div style="width: 320px; aspect-ratio: 1.6;" data-expected-height=200></div>
</body>
